Doctor Section, Services Section added
After figuring out the services section to have animations and such, I need to go through and optimize the site for mobile devices. 
That will need to be a jquery hamburger menu, some media queries for the services section and the doctor section, and will need to optimize the information section for each of the services once they are clicked on.